+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## XX.XX.XX
+
+* Metric override is added and accessible through "CountlyConfig.SetMetricOverride()"
 * "cpuId" and "multipleFields" device id generation methods are deprecated from Net Framework 3.5 and 4.5 versions
 
 ## 23.12.0